pyproject.toml
[project]
name = "utility-mcp-server"
version = "0.1.0"
description = "MCP server for EV utility function calculations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "upstash-redis>=1.1.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
